Removed tons of warnings (again ... :-/)
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@18608 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 20a932064828f15ca143503b8f60d1ef2347aafb

+++ common/mep/src/de/uniol/inf/is/odysseus/mep/functions/StoreFunction.java
@@ -57,7 +57,7 @@ public class StoreFunction extends AbstractFunction<Double> {
         Objects.requireNonNull(index);
         IAssociativeStorage<Tuple<?>> store = AssociativeStorageManager.get(name);
         if (store != null) {
-            return (Double) store.get(path, index);
+            return store.get(path, index);
         }
         else {
             return 0.0;
